rename main fixture file
